+++ px4parser.py
@@ -1,386 +1,8 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import print_function
-
-"""Dump binary log generated by PX4's sdlog2 or APM as TXT
-    
-Usage: python px4parser <log.bin>  [-e] [-d delimiter] [-def] [-eng] [-rus] [-m MSG[.field1,field2,...]] [-f file.txt]
-
-	-e	    Recover from errors.
-    
-    -d      Use "delimiter" in file. Default is TAB.
-    
-    -def    Use default namespace.
-
-    -eng    Use customized English namespace.
-
-    -rus    Use customized Russian namespace.
-
-    -c      Use constant clock.
-    
-    -m      MSG[.field1,field2,...]
-            Dump only messages of specified type, and only specified fields.
-            Multiple -m options allowed.
-
-    -f      Set output file name.
-"""
-
-__author__  = "Roman " + "broken_cursor" + " Shvindt. Based on sdlog2_dump"
-__version__ = "v0.6"
-
 import struct, sys
-
-if sys.hexversion >= 0x030000F0:
-    runningPython3 = True
-    def _parseCString(cstr):
-        return str(cstr, 'ascii').split('\0')[0]
-else:
-    runningPython3 = False
-    def _parseCString(cstr):
-        return str(cstr).split('\0')[0]
-
-class SDLog2Parser:
-    BLOCK_SIZE = 8192
-    MSG_HEADER_LEN = 3
-    MSG_HEAD1 = 0xA3
-    MSG_HEAD2 = 0x95
-    MSG_FORMAT_PACKET_LEN = 89
-    MSG_FORMAT_STRUCT = "BB4s16s64s"
-    MSG_TYPE_FORMAT = 0x80
-    FORMAT_TO_STRUCT = {
-        "a": ("a", None),
-        "b": ("b", None),
-        "B": ("B", None),
-        "h": ("h", None),
-        "H": ("H", None),
-        "i": ("i", None),
-        "I": ("I", None),
-        "f": ("f", None),
-        "d": ("d", None),
-        "n": ("4s", None),
-        "N": ("16s", None),
-        "Z": ("64s", None),
-        "c": ("h", 0.01),
-        "C": ("H", 0.01),
-        "e": ("i", 0.01),
-        "E": ("I", 0.01),
-        "L": ("i", 0.0000001),
-        "M": ("b", None),
-        "q": ("q", None),
-        "Q": ("Q", None),
-    }
-    __delim_char = "\t"             # Character to put in between columns
-    __null_char = ""                # Charecter to put if there is no data in column 
-    __msg_filter = []               # List of messages to output
-    __msg_ignore = []               # List of messages to ignore during processing
-    __msg_id_ignore = []            # IDs of messages to ignore during processing. Fills automaticaly from data in __msg_ignore
-    __time_msg = None               # Name of the message that contains time
-    __time_msg_id = 0               # Time message ID. Fills automaticaly
-    __status_msg = None             # Name of the status message
-    __status_msg_id = 0             # ID of the status message. Fills automaticaly
-    __debug_out = False             # Debug output flag
-    __correct_errors = False        # Error corrcetion flag
-    __file = None                   # Output file name
-    __namespace = {}                # Columns names to print instead of __txt_columns
-    __prev_data = []                
-    __next_data = []
-    __constant_clock = False        # Constant message output clock flag
-    msg_count = 0                   
-    
-    def __init__(self):
-        return
-
-    
-    def reset(self):                # Reset variables default
-        self.__msg_descrs = {}      # Message descriptions by message type map
-        self.__msg_labels = {}      # Message labels by message name map
-        self.__msg_names = []       # Message names in the same order as FORMAT messages
-        self.__buffer = bytearray() # Buffer for input binary data
-        self.__pointer = 0          # Read pointer in buffer
-        self.__txt_columns = []     # CSV file columns in correct order in format "MSG.label"
-        self.__txt_data = {}        # Current values for all columns
-        self.__txt_updated = False  # Is updated flag
-        self.__msg_filter_map = {}  # filter in form of map, with '*" expanded to full list of fields
-        self.__prev_data = []       # Previous data buffer 
-        self.__next_data = []       # Next data buffer
-        self.msg_count = 0          # Printed messages count
-    
-    def setNamespace(self, namespace):      # Set a custom namespace to print instead of __txt_columns
-        self.__namespace = namespace
-
-    def setDelimiterChar(self, csv_delim):  # Set a char to print in between columns
-        self.__delim_char = csv_delim
-    
-    def setNullChar(self, csv_null):        # Set a char to print if there is no data in the column
-        self.__null_char = csv_null
-    
-    def setMsgFilter(self, msg_filter):     # Set a filter of messages to output
-        self.__msg_filter = msg_filter
-    
-    def setTimeMsg(self, time_msg):         # Set time message name
-        self.__time_msg = time_msg
-
-    def setDataMsg(self, status_msg):       # Set a status message name
-        self.__status_msg = status_msg
-    
-    def setDebugOut(self, debug_out):       # Set a debug output flag
-        self.__debug_out = debug_out
-
-    def setCorrectErrors(self, correct_errors): # Set an error correction flag
-        self.__correct_errors = correct_errors
-
-    def setConstantClock(self, constant_clock): # Set a constant clock flag
-        self.__constant_clock = constant_clock
-
-    def setMsgIgnore(self, msg_ignore):     # Set a list of messages to ignore during processing
-        self.__msg_ignore = msg_ignore
-
-    def setFileName(self, file_name):       # Set an output file's name
-    	self.__file_name = file_name
-    	if file_name != None:
-    		self.__file = open(file_name, 'w+')
-    	else:
-    		self.__file = None
-
-    
-    def process(self, fn): # Main function
-        self.reset()
-        if self.__debug_out:
-            # init __msg_filter_map
-            for msg_name, show_fields in self.__msg_filter:
-                self.__msg_filter_map[msg_name] = show_fields
-        first_data_msg = True
-        f = open(fn, "rb")
-        bytes_read = 0
-        while True:
-            chunk = f.read(self.BLOCK_SIZE)
-            if len(chunk) == 0:
-                break
-            self.__buffer = self.__buffer[self.__pointer:] + chunk
-            self.__pointer = 0
-            while self.__bytesLeft() >= self.MSG_HEADER_LEN:
-                head1 = self.__buffer[self.__pointer]
-                head2 = self.__buffer[self.__pointer+1]
-                if (head1 != self.MSG_HEAD1 or head2 != self.MSG_HEAD2):
-                    if self.__correct_errors:
-                        self.__pointer += 1
-                        continue
-                    else:
-                        raise Exception("Invalid header at %i (0x%X): %02X %02X, must be %02X %02X" % (bytes_read + self.__pointer, bytes_read + self.__pointer, head1, head2, self.MSG_HEAD1, self.MSG_HEAD2))
-                msg_type = self.__buffer[self.__pointer+2]
-                if msg_type == self.MSG_TYPE_FORMAT:
-                    # parse FORMAT message
-                    if self.__bytesLeft() < self.MSG_FORMAT_PACKET_LEN:
-                        break
-                    self.__parseMsgDescr()
-                else:
-                    # parse data message
-                    msg_descr = self.__msg_descrs[msg_type]
-                    if msg_descr == None:
-                        raise Exception("Unknown msg type: %i" % msg_type)
-                    msg_length = msg_descr[0]
-                    if self.__bytesLeft() < msg_length:
-                        break
-                    if first_data_msg:
-                        # build TXT columns and init data map
-                        if not self.__debug_out:
-                            self.__initOutput()
-                        first_data_msg = False
-                    self.__parseMsg(msg_descr)
-            bytes_read += self.__pointer
-            if not self.__debug_out and self.__time_msg != None and self.__txt_updated:
-                self.__processData()
-        f.close()
-    
-    def __bytesLeft(self):              # Get an amout of bytes left 
-        return len(self.__buffer) - self.__pointer
-    
-    def __filterMsg(self, msg_name):    # Process filtering
-        show_fields = "*"
-        if len(self.__msg_filter_map) > 0:
-            show_fields = self.__msg_filter_map.get(msg_name)
-        return show_fields
-    
-    def __initOutput(self):             # Init programs output
-        if len(self.__msg_filter) == 0: # If filter is impty, fill it with eptyness sings
-            for msg_name in self.__msg_names:
-                self.__msg_filter.append((msg_name, "*"))
-
-        for msg_name, show_fields in self.__msg_filter:         # Fill in __txt_columns and __txt_data in accrodig to the __msg_filter
-            if show_fields == "*":
-                show_fields = self.__msg_labels.get(msg_name, [])
-            self.__msg_filter_map[msg_name] = show_fields
-            for field in show_fields:
-                full_label = msg_name + "_" + field
-                self.__txt_columns.append(full_label)
-                self.__txt_data[full_label] = None
-
-        for i in self.__txt_columns:                            # Fill in __msg_ignore_id in accroding to the __msg_ignore
-            if i in self.__msg_ignore:                          # If message name is present in __msg_ignore
-                self.__msg_id_ignore.append(self.__txt_columns.index(i)) # Put it's ID from __txt_columns to __msg_ignore_id
-        self.__status_msg_id = self.__txt_columns.index(self.__status_msg) # Get status message id from __txt_columns and put it in __status_msg_id
-        self.__time_msg_id = self.__txt_columns.index(self.__time_msg)     # Get time message id from __txt_columns and put it in __time_msg_id
-
-        headers = [] # Create a headers list
-        if self.__namespace: # If __namespace is not empty (i.e use __namespace instead of __txt_columns in output)
-            for column in self.__txt_columns: # Check every column
-                if column in self.__namespace: # If it has a custom name, use it
-                    headers.append(self.__namespace[column])
-                else:
-                    headers.append(column) # If there doest have a custom name, use default from __txt_columns 
-        else:
-            headers = self.__txt_columns # If __namesapce is empty, use default columns headers (i.e. __txt_columns)
-
-        if self.__file != None: # Output headers
-            print(self.__delim_char.join(headers), file=self.__file)
-        else:
-            print(self.__delim_char.join(headers))
-    
-
-    def __processData(self): # Process raw data 
-        data = []
-        for full_label in self.__txt_columns: # Fill in data accordigly to __txt_columns
-            v = self.__txt_data[full_label]   # Get single string from data dictionary
-            if v == None: # If string is empty
-                v = self.__null_char # Put null character in
-            else:
-                v = str(v) # If there is some data, convert it to a string
-            data.append(v) # Add new string to data list
-
-        if self.__constant_clock: # If constant clock flag is true
-            if not self.__next_data: # If __next_data is empty
-                self.__next_data = data[:] # Fill it with current data
-                return 0 # Exit the function
-            else: # If __next_data contains something 
-                curr_data = self.__next_data[:] # Put data from __next_data to curr_data
-                prev_data = curr_data[:] # Make a bakup to put in __prev_data later
-                self.__next_data = data[:] # Update __next_data 
-                if self.__prev_data: # if __prev_data is not empty
-                    time_diff = ((int(curr_data[self.__time_msg_id]) - int(self.__prev_data[self.__time_msg_id])) // 1000) # Calcuate a time difference between previous message and current message 
-                    to_round_time = float((100 - (time_diff%100)) * 0.0001) # Calculate a multiplier based on time difference for __curr_data to reach a round digit time  
-                    extra_msg_count = time_diff // 100 # Calculate an amount of extra messages for correct extrapolation
-                    for count in range(1, extra_msg_count + 1): # For every extra message 
-                        if count == 1 and to_round_time != 1.0: # If it's first extra message and there's a time is not a round digit. This if statement calculates new beginig point with round time if original doesnt have a round time. This is necessary to maintain a constant clock
-                            for i in range(len(curr_data)): # Multiply every string in list by to_round_time multiplier
-                                if i == self.__time_msg_id or curr_data[i] == self.__prev_data[i] or i in self.__msg_id_ignore: # If time message is being processed or curr_data at id [i] and __prev_data at id[i] is equal 
-                                    continue # Skip this iteration (id)
-                                if self.__next_data[i] > curr_data[i]: # If __next_data is greater than curr_data
-                                    curr_data[i] = str(float(curr_data[i]) + (float(curr_data[i]) * to_round_time)) # Convert string to float, add (curr_data at id i * to_round_time) to curr_data, convert everything back to string and put into curr_data at id i
-                                else:
-                                    curr_data[i] = str(float(curr_data[i]) - (float(curr_data[i]) * to_round_time)) # Convert string to float, substract (curr_data at id i * to_round_time) from curr_data, convert everything back to string and put into curr_data at id i
-                            curr_data[self.__time_msg_id] = str(self.msg_count * 100) # Calculate current time by multiplier amount of messages by clock time (100 ms)
-                            self.__printData(curr_data) # Print curr_data
-                            self.msg_count += 1 # Add extra message
-                        tmp = curr_data[:] # Create a temporary dada list
-                        for id in range(len(curr_data)): # Extrapolate data 
-                            if id in self.__msg_id_ignore or self.__next_data[id] == curr_data[id]: # If message in ignore list or equal to __next_data
-                                continue
-                            elif self.__next_data[id] > curr_data[id]: # If curr_data[id] less than __next_data
-                                tmp[id] = str(float(curr_data[id]) + (((float(self.__next_data[id]) - float(curr_data[id])) / extra_msg_count) * count)) # Add extrapolation
-                            else:  # If curr_data[id] greater than __next_data
-                                tmp[id] = str(float(curr_data[id]) - (((float(curr_data[id]) - float(self.__next_data[id])) / extra_msg_count) * count)) # Substract extrapolation
-                        tmp[self.__time_msg_id] = str(self.msg_count * 100) # Calculate current time by multiplying amount of messages by clock time (100 ms)
-                        self.__printData(tmp)  # Print data              
-                        self.msg_count += 1 # Add extra message
-                else: # If __prev_data empty
-                    curr_data[self.__time_msg_id] = str(self.msg_count * 100) # Calculate current time by multiplying amount of messages by clock time (100 ms)
-                    self.__printData(curr_data) # Print data
-                    self.msg_count += 1 # Add extra message
-                self.__prev_data = prev_data[:] # Update __prev_data from dedicated buffer
-        else: # if constant clock is false
-           self.__printData(data) # Print data
-
-    def __parseMsgDescr(self): # Get message description
-        # How does it work? idunno
-        if runningPython3:
-            data = struct.unpack(self.MSG_FORMAT_STRUCT, self.__buffer[self.__pointer + 3 : self.__pointer + self.MSG_FORMAT_PACKET_LEN])
-        else:
-            data = struct.unpack(self.MSG_FORMAT_STRUCT, str(self.__buffer[self.__pointer + 3 : self.__pointer + self.MSG_FORMAT_PACKET_LEN]))
-        msg_type = data[0]
-        if msg_type != self.MSG_TYPE_FORMAT:
-            msg_length = data[1]
-            msg_name = _parseCString(data[2])
-            msg_format = _parseCString(data[3])
-            msg_labels = _parseCString(data[4]).split(",")
-            # Convert msg_format to struct.unpack format string
-            msg_struct = ""
-            msg_mults = []
-            for c in msg_format:
-                try:
-                    f = self.FORMAT_TO_STRUCT[c]
-                    msg_struct += f[0]
-                    msg_mults.append(f[1])
-                except KeyError as e:
-                    raise Exception("Unsupported format char: %s in message %s (%i)" % (c, msg_name, msg_type))
-            msg_struct = "<" + msg_struct   # force little-endian
-            self.__msg_descrs[msg_type] = (msg_length, msg_name, msg_format, msg_labels, msg_struct, msg_mults)
-            self.__msg_labels[msg_name] = msg_labels
-            self.__msg_names.append(msg_name)
-            if self.__debug_out:
-                if self.__filterMsg(msg_name) != None:
-                    print("MSG FORMAT: type = %i, length = %i, name = %s, format = %s, labels = %s, struct = %s, mults = %s" % (
-                                msg_type, msg_length, msg_name, msg_format, str(msg_labels), msg_struct, msg_mults))
-        self.__pointer += self.MSG_FORMAT_PACKET_LEN
-    
-    def __parseMsg(self, msg_descr): # Get raw data from file
-        msg_length, msg_name, msg_format, msg_labels, msg_struct, msg_mults = msg_descr  #Disassemble msg_descr    
-        if not self.__debug_out and self.__time_msg != None and msg_name == self.__time_msg and self.__txt_updated: 
-            self.__processData()
-            self.__txt_updated = False
-        show_fields = self.__filterMsg(msg_name) # Get data from filter
-
-        if (show_fields != None): # If filter is not empty
-            if runningPython3:
-                data = list(struct.unpack(msg_struct, self.__buffer[self.__pointer+self.MSG_HEADER_LEN:self.__pointer+msg_length])) # Parse data from file to list
-            else:
-                data = list(struct.unpack(msg_struct, str(self.__buffer[self.__pointer+self.MSG_HEADER_LEN:self.__pointer+msg_length]))) # Parse data from file to list
-
-            for i in range(len(data)): # For every object in list
-                if type(data[i]) is str: # if data type is string
-                    data[i] = _parseCString(data[i]) # Convet to readable string
-                m = msg_mults[i] # Get multiplier from structure
-                if m != None: # If there is a multiplier
-                    data[i] = data[i] * m # Multiply
-
-            if self.__debug_out: # If debug is enabled
-                s = []
-                for i in range(len(data)):
-                    label = msg_labels[i]
-                    if show_fields == "*" or label in show_fields:
-                        s.append(label + "=" + str(data[i]))
-                            
-                print("MSG %s: %s" % (msg_name, ", ".join(s)))
-
-            else: # If debug is disabled
-                # update data buffer
-                for i in range(len(data)): # For every column
-                    label = msg_labels[i] # Get label
-                    if label in show_fields: # If label is in filter
-                        self.__txt_data[msg_name + "_" + label] = data[i] # Add parsed raw data to __txt_data
-                        if self.__time_msg != None and msg_name != self.__time_msg:
-                            self.__txt_updated = True
-                if self.__time_msg == None:
-                    self.__processData()
+from pxparser import Pxparser 
 
 
-        self.__pointer += msg_length # move pointer
-        
-    def __printData(self, data): # Print to output
-        for id in range(len(data)): # For every mes
-            if id == self.__status_msg_id: # if processing status message
-                data[self.__status_msg_id] = data[self.__status_msg_id].strip("'").lstrip("b'").strip('\\x00') # clear up all garbage
-                continue # Skip everything else
 
-            if len(data[id]) > 10 and id != self.__time_msg_id: # If message is longer than 10 characters
-                data[id] = data[id][:8] # Trim to 8 characters
-            elif len(data[id]) < 8: # if message is shorter than 8 characters
-                data[id] += '\t' # add tab
-
-        if self.__file != None: # Print data
-            print(self.__delim_char.join(data), file=self.__file)
-        else:
-            print(self.__delim_char.join(data))
-       
 
 def _main():
     if len(sys.argv) < 2:
@@ -390,7 +12,7 @@ def _main():
         print("\t-d\tUse \"delimiter\" in output. Default is TAB.\n")
         print("\t-def\tUse default namespace.\n")
         print("\t-eng\tUse customized English namespace.\n")
-        print("\t-eng\tUse customized Russian namespace.\n")
+        print("\t-rus\tUse customized Russian namespace.\n")
         print("\t-c\tUse constant clock.\n")
         print("\t-m MSG[.field1,field2,...]\n\t\tDump only messages of specified type, and only specified fields.\n\t\tMultiple -m options allowed.\n")
         print("\t-t\tSpecify TIME message name to group data messages by time\n")
@@ -448,23 +70,23 @@ def _main():
 
     if delim_char == "\\t":
         delim_char = "\t"
-    parser = SDLog2Parser()
-    parser.setDelimiterChar(delim_char)
-    parser.setNullChar(null_char)
-    parser.setMsgFilter(msg_filter)
-    parser.setTimeMsg(time_msg)
-    parser.setDataMsg(data_msg)
-    parser.setFileName(file_name)
-    parser.setDebugOut(debug_out)
-    parser.setConstantClock(constant_clock)
-    parser.setCorrectErrors(correct_errors)
+    parser = Pxparser()
+    parser.set_delimiter_char(delim_char)
+    parser.set_null_char(null_char)
+    parser.set_msg_filter(msg_filter)
+    parser.set_time_msg(time_msg)
+    parser.set_data_msg(data_msg)
+    parser.set_fileanme(file_name)
+    parser.set_debug_flag(debug_out)
+    parser.set_constant_clock_flag(constant_clock)
+    parser.set_error_corr_flag(correct_errors)
     if custom_namespace == "def":
-        parser.setNamespace(default_namespace)
+        parser.set_namespace(default_namespace)
     elif custom_namespace == "rus":
-        parser.setNamespace(custom_rus_namespace)
+        parser.set_namespace(custom_rus_namespace)
     elif custom_namespace == "eng":
-        parser.setNamespace(custom_eng_namespace)
-    parser.setMsgIgnore(msg_ignore)
+        parser.set_namespace(custom_eng_namespace)
+    parser.set_msg_ignore(msg_ignore)
     parser.process(fn)
     print("Done\nParsed " + str(parser.msg_count + 1) + " lines")
